fix: label as float64 due to problems with pearson coef metrics

+++ mammaprint/datamodule/datasets/mil.py
@@ -28,6 +28,6 @@ class MILDataset(BaseDataset):
         # Check if any tile in `sample` has `self.label` not 0.0
         label_value = next((tile.get(self.label) for tile in sample if tile.get(self.label, 0.0) != 0.0), 0.0)
         # logging.info(f"label value found: {label_value}")
-        label = torch.tensor([label_value])
+        label = torch.tensor([label_value], dtype=torch.float64)
 
         return images_tensor, label, sample
